+++ transport_handbook/src/transport_catalogue.h
@@ -13,17 +13,17 @@ class TransportCatalogue
 
 	double RoutePathSizeNaive(const std::vector<std::string>& stops) const;
 
-	double RoutePathSize(const std::vector<std::string>& stops) const;
+	size_t RoutePathSize(const std::vector<std::string>& stops) const;
 
-	std::optional<std::tuple<size_t, size_t, double, std::vector<std::string>, double>> GetRouteInfo(
+	std::optional<std::tuple<size_t, size_t, size_t, std::vector<std::string>, double>> GetRouteInfo(
 		const std::string& bus) const;
 
 	std::optional<std::set<std::string>> GetBusInfo(const std::string& stop) const;
 
-	double GetDistanceBetweenStop(const std::string& stop_l, const std::string& stop_r) const;
+	size_t GetDistanceBetweenStop(const std::string& stop_l, const std::string& stop_r) const;
 
   private:
-	std::unordered_map<std::string, std::unordered_map<std::string, double>> distance_between_stops_;
+	std::unordered_map<std::string, std::unordered_map<std::string, size_t>> distance_between_stops_;
 	std::unordered_map<std::string, Coordinates> stops_;
 	std::unordered_map<std::string, std::vector<std::string>> bus_to_stops_;
 	std::unordered_map<std::string, std::set<std::string>> stop_to_bus_;
